requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.66
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.4.1
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
mutagen==1.47.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
ollama==0.5.1
openai-whisper @ git+https://github.com/openai/whisper.git@25639fc17ddc013d56c594bfbf7644f2185fad84
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pycryptodomex==3.21.0
pydantic==2.11.7
pydantic_core==2.33.2
pytube==15.0.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sympy==1.13.1
tenacity==9.1.2
tiktoken==0.8.0
torch==2.5.0
tqdm==4.66.5
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
xxhash==3.5.0
youtube-transcript-api==0.6.2
yt-dlp==2024.10.7
zstandard==0.23.0
